Fix "buildartifactsurl" parameter for cloud storage bucket to Autotest Tast runner.
This parameter is required to run Tast tests from Moblab.

The current implementation assume test that which is not what
moblab uses.

If there is no build command passed in use the cros-version label
on the host to build the URL for the build artifacts.

BUG=chromium:1069786
TEST=local build and run on moblab

Change-Id: Ic5b8871bb2c0383fabfe7ca145decbe51f575cf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2345071
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Alex Bergman <abergman@google.com>
Commit-Queue: Keith Haddow <haddowk@chromium.org>
Auto-Submit: Keith Haddow <haddowk@chromium.org>