firmware_test: ignore the warm_reset remap issue
warm_reset is remapped to the ccd control on kukui when dual v4 is
enabled. That interface may not be available if the test has disabled
ccd before calling try to bring dut up. Ignore the interface unavailable
error until the remap issue is resolved. Turning cold_reset off is
enough on kukui and this issue doesn't happen on other boards. Ignoring
this error is harmless.

BUG=b:159338538
TEST=run faft-cr50 on jacuzzi

Change-Id: Ib66ee8ce7d5fcdaa63592fa417a980e5a89b1762
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2347389
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>