servod: Use per-thread proxy server
httplib is not multi-threaded, so create a new proxy server connection
for each separate thread that requests it (if not using a ssh tunnel).

Some tests use threading.Timer to create asynchronous requests, and a
new thread is created for each of these objects, requiring a separate
xmlrpc connection for each thread.

BUG=b:163727019
TEST=Run firmware_ECPowerButton and confirm test works

Change-Id: I8e5b8ee94d3e48b9de009799d6a2ae4a63034b2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2353649
Tested-by: Andrew McRae <amcrae@chromium.org>
Commit-Queue: Andrew McRae <amcrae@chromium.org>
Reviewed-by: Andrew McRae <amcrae@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>