power_VideoPlayback: 30s / 4 loops instead of 2m video
2 minutes VP8 4K 60 fps video size is about 550 MiB. Adding
that with Chrome video player and autotests overhead can
sometime makes DUT with 2GB total memory runs out of memory
randomly in the middle of the test which make this test flaky.

This CL changes the video to 30 seconds one and loop it 4 times
instead to mitigate this.

BUG=b:158851008
TEST=Ran the test

Change-Id: Ia5a4650a319f2ffaf1e460c9787c3cdc72dedd00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2351760
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>
Auto-Submit: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>