FAFT: Added a delay between retries when checking power state
Added a new parameter to the wait_power_state() API that allows the
user to specify a delay between retries when checking the device power
state. The default value is zero to ensure the timing is the same for
other tests that use the wait_power_state() function.

BUG=b:159944771
TEST=Verified test passes on eve and kohaku

Change-Id: I8b0d04ea3d8a7eb8ee6fbcf0d68d2d4b81c917af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2354934
Tested-by: Brent Peterson <brentpeterson@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Brent Peterson <brentpeterson@chromium.org>