bluetooth: Add delay after connect in role tests
I have found that some tests are flaky on kernel 4.4 if we connect to a
peer and immediately connect again to another one. In this case,
sometimes the first device does not get assigned an input device,
causing tests to fail. This test makes sure to delay after connecting
the first time to improve test stability.

BUG=b:164131633
TEST=15 iterations of le_role_hid_during_sender test

Change-Id: Iabfe2a60b87c702d934733c896cab07824cdf782
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2354768
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>