bluetooth: Increase receive time in role tests
Our raspberry pi peer has poor support for advertising. The fastest
advertising interval is >1 second and can not be configured. For this
reason, it can take longer to receive an advertisement in our controller
role tests. This change increases the time to receive to reduce
flakiness here.

This is not a real issue, it is just a limitation of our test setup.

BUG=b:164164169
TEST=None since the issue is very rare

Change-Id: I2a37816371bf6532de31914152846fb4e5b0f8e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2355109
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>