Add AU test for periodic update check.
In CL:2333308 a pref was added to enable periodic checks on test
images. This is a test for that functionality. We first check for a
periodic check that returns no update, then another that returns an
update and verifies that the update completes successfully.

BUG=chromium:953471
TEST=autoupdate_Periodic

Change-Id: I42a39489f272549c01b1925b4da0613c55f07135
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2354936
Commit-Queue: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>