bluetooth: a2dp: generate and play raw file
The cras_test_client only plays back raw files. We should use
raw files for playback instead of wav files. Otherwise, the
wav header would be mistaken as audio data.

BUG=b:164918095
TEST=Run bluetooth_AdapterAUSanity.au_a2dp_test.
The recorded wav file on the Pi should not have a spike in the
beginning of the recorded streams.

Change-Id: I4a1e739ec3cb45438f7b4dbdb127304e5ced0c54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2359093
Reviewed-by: Michael Sun <michaelfsun@google.com>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>