autoupdater.py: Temporarily remove the tast bundle files when provisioning
It seems like these files are downloaded by tasts themselves and are not
cleaned up properly so they are wasting a lot of space and causing
provisioning script to fail. This is temporary until we figure out why
tast tests are not doing proper clean up.

BUG=b:165024723
TEST=None

Change-Id: Ib33fefd6acc4f70ca9bd1fa6c8bda31fd2fb0c64
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2360536
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>