mode_switcher: Add delay_powerinfo_stable config and handling
Delay checking the powerinfo via the EC until it is stable.
Uses a new platform config item to delay checking the power state
when waiting for the DUT to come up.
Also fix the firmware_ECPowerButton test so that it incorporates the
startup delay when waiting for the power state to change.

BUG=b:165124883
TEST=Confirm that test passes with Puff.

Cq-Depend: 2359878
Change-Id: I3428472bbbd67b3032abea46424eae951c85469a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2359100
Tested-by: Andrew McRae <amcrae@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>