bluetooth: Ignore some failures in do_suspend
The peer wake tests are failing even though the peer is correctly waking
up the system. The problem seems to be that the wakeup reason isn't
correctly being propagated to a level where powerd is reading and
logging it. Until we fix that or use some other source for determining
BT wakeup, just disable this and depend on the wakeup timeout for test
success.

BUG=b:160803597
TEST=Ran peer wake tests with hacks to disable BT wake to make sure
     timeout catches failures

Change-Id: I25a08d2605e19cf367ce841dbe68341f22c7715f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2365055
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>