faft: Split cbfs add/remove, add extension args
In anticipation of creating tests that manipulate CBFS only in one
slot, this change adds an argument to specify the slot/region to
act on.  It also adds the ability to pass in extensions for the bin
and hash files, to avoid having to guess

It also splits the add/remove logic, because it simplifies the
extension arg handling.

The cbfs_extract_chip method also now reports the full paths of the
extracted files, in a dict for compatibility with boolean checks.

BUG=b:165442321
TEST=Run firmware_FAFTRPC.updater

Change-Id: I683f0856d5bfea073ac39896baaf6436a70bf963
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2363866
Commit-Queue: Dana Goyette <dgoyette@chromium.org>
Tested-by: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>