bluetooth: hfp: generate and play raw files
The cras_test_client only plays back raw files. We should use
raw files for playback instead of wav files. Otherwise, the
wav header would be mistaken as audio data and results in
an initial spike in the recorded stream.

BUG=b:165480358
TEST=Run the following tests
  - bluetooth_AdapterAUSanity.au_hfp_wbs_dut_as_source_test
  - bluetooth_AdapterAUSanity.au_hfp_nbs_dut_as_source_test
The recorded wav files on the Pi should not have a spike in the
beginning of the recorded streams.

Change-Id: Idf878248d4cbae56201305d2b7a3857a7e0adc84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2363511
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>