bluetooth: 2 VISQOL score tests for WBS
Two tests that use visqol to score a series of files as described in
bluetooth_audio_test_data.py using the VISQOL functions. The tests pass
if every file meets the minimum threshold described in its dictionary
which for now is set to 0 due to issues in the dut as sink direction.

BUG=b:162273392
TEST=bluetooth_AdapterAUSanity

Change-Id: If7fc7eace33efef210cb3c621aa046c0fd291191
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2364482
Commit-Queue: Kshitij Shah <tij@google.com>
Tested-by: Kshitij Shah <tij@google.com>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>