[Autotest Infra] Quick fix to autotest.verify_machine
Changing the test -e to an echo, as the exit status would be covered by
the umounts. Additionally changing && to ; as && does not continue if
the first command has a non-zero exit status. Testing a negative case
this time and it failed, correctly :)

BUG=b:163878321
test=Dummy_Pass, autotest_unittest

Change-Id: Ib42bc452d95c1c0bdfd853c6de6af1a70e40916d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2373208
Tested-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>