bluetooth: switch tester initialization failures to throw TestNA
We have a long-time peer flake where the bluetooth_tester emulated
device fails to disable Secure Connections at the beginning of the test.
When this occurs, the test fails with a TestFail. Since this isn't a DUT
issue, we are switching it to throw TestNA instead so we don't clutter
the dashboard while we find a workaround.

BUG=b:146149884
TEST=Ran SDP tests and forced peer failure to confirm TestNA is thrown

Change-Id: I3877034747a0932f0c4c0f885ffba367a114c4cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2378718
Tested-by: Daniel Winkler <danielwinkler@google.com>
Auto-Submit: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>