bluetooth: select the bluetooth device as audio output node
To run Bluetooth audio tests with the disabled ui process, we need to
explicitly select the paired Bluetooth device as the outpout node so
that we do not reply on chrome to do this.

BUG=b:166715901, b:166308500
TEST=Run bluetooth_AdapterAUSanity tests with
the following CL that disable ui
bluetooth: Disable ui for BT tests (http://crrev.com/c/2370023)

The tests should pass.

Change-Id: I7322c4e22d437b3562dd6af19b496a7f2cd4f8aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2379496
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Michael Sun <michaelfsun@google.com>