Remove the crash-after-update part of autoupdate_Interruptions.
The crash after update only failed once in R29 and the bug was marked
WontFix. The automated test is not even doing what the manual test
wanted it to do: check the settings page for a reboot button after the
crash. So we can just remove the crash part since it is not really
related to interrupting an update anyway.

I'll remove the client test that does the crashing. It also logged in
with a new user. I've moved that functionality to the other client test
used here.

BUG=chromium:1087347
TEST=autoupdate_Interruptions

Change-Id: I1a44645c88d457e25926017fad3136ceb3283db2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2405597
Commit-Queue: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>