power_suspendStress: disable modemfwd in FSI suspend stress test
On boards with both a LTE and a non-LTE SKU, at startup, the modemfwd
modem firmware updater is trying to update the (possibly non-existing)
modem firmware and timeout'ing after 2 seconds if it is absent.
In that case, modemfwd is taking the suspend lock for 2 seconds and
making the suspend stress test fail. As a temporary measure, disable the
modemfwd daemon when doing the FSI testing. Later the modemfwd daemon
will be aware of the existence (or not) of the modem.

BUG=b:164255562
TEST=test_that --board=drallion $DUT_IP f:.*power_SuspendStress/control.bareFSI

Change-Id: I4345646540b13f0dbf0ae07e7a36f42a9e149a94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2405585
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>