bluetooth: add library support for A2DP LONG test
This patch adds the underlying support for a new A2DP test which has
a longer audio stream for testing. The recorded stream on Pi is
divided into chunks on the fly which are sent to the DUT to verify
the primary frequencies.

BUG=b:161953821
TEST=None

Cq-Depend: chromium:2373975
Change-Id: Ic236b02d268a2da2a12ce4feee1f0c73cbcca2b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2374825
Reviewed-by: Michael Sun <michaelfsun@google.com>
Reviewed-by: Yu Liu <yudiliu@google.com>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>