Remove unused assert_users_on_whitelist()
Prior to this CL, the only code which called assert_users_on_whitelist()
was a conditional block in compare_policy_response(), activated iff
compare_policy_response() was called with the optional "whitelist"
param.

But, grep shows that compare_policy_response() is never called with
whitelist, so it is dead code, and should be removed.

BUG=b:159917945
TEST=None

Change-Id: Ide7cf4bb85baf26e05f3b7ed3b3536008898d819
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2410566
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>