autotest: Add timeout to _ping_check_status
In Bluetooth MTBF tests, we encountered a scenario where a ping command
that should have completed in 1s (with a failure) completed almost
8 seconds later with success. Either ping failed to enforce the deadline
(set via -w1) or the subprocess call got stuck somewhere.

Here is an example log of the issue:

09/12 23:43:15.056 DEBUG| utils:0219| Running 'ping chromeos15-row5-rack5-host6 -w1 -c1'
09/12 23:43:23.067 DEBUG| utils:1993| [rc=0] 1 packets transmitted, 1 received, 0% packet loss, time 0ms; rtt min/avg/max/mdev = 2.923/2.923/2.923/0.000 ms

To enforce a more consistent ping check, also use the timeout parameter
when running the ping request. Either the ping will finish quickly (less
than deadline), finish at the deadline (which will likely be before the
timeout) or after the timeout (at which point the subprocess will be
killed).

BUG=b:168428728
TEST=Ran a bluetooth specific suspend/resume stress test that uses
     cros_hosts.test_wait_for_sleep and passed 50 iterations without the
     issue.

Change-Id: Ib1b0d6359b21c0dff792914f6904067d4abfc390
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2411453
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Yu Liu <yudiliu@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>