cheets_CTS_Instant: Split into multiple jobs.
For two reasons:
 * Reduce a single job runtime (<2h is desired)
 * To decouple flaky tests (CtsAccessibility and CtsAutoFill) from
   others, so that they don't affect other tests badly.

Tricky part is that cts-instant does not support the usual
'--include-filter' option, because internally it is implemented
as a list of include-filters for modules supporting "instant app"
testing. Adding addition include-filter from the command line app
is no-opp because the full set is already 'include-filter'ed.

Instead, exclude-filter works. Considering that the whole module
count is not too big, we simulated include-fitlers by excluding
all the modules NOT to be included.

BUG=b:150348473
TEST=cheets_CTS_Instant.9.0_r13.arm.all.CtsSampleDeviceTestCases_-_\
  CtsWindowManagerDeviceTestCases

Change-Id: Ide2b94d7c0834613174ca15ccbbd7dd6ed5989c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2417732
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Jiyoun Ha <jiyounha@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>