bluetooth: make a copy of test data for distinct arguments
For the A2DP test, it is flexible to specify arguments for the
test profile. Each test can have its own arguments. Hence,
we need to make a copy of the test data and discard the local copy
when the test is completed.

BUG=b:169110093
TEST=Execute A2DP tests a few more times in bluetooth_AdapterAUSanit
and they should all pass.

Change-Id: I46d18e34737f11f35dd72e277708db2fc9680834
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2422108
Reviewed-by: Michael Sun <michaelfsun@google.com>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>