Pass disable-hid-detection-on-oobe flag during OOBE AU tests.
Recently OOBE was updated to better detect when there is no keyboard
attached to a device. This meant that on chromebox/chromebases that
don't have keyboards attached in the lab the OOBE AU tests began to
fail.

This CL adds the flag to chrome when the OOBE AU tests start.

BUG=chromium:1127536
TEST=autoupdate_ForcedOOBEUpdate before and after CL.

Change-Id: I3b0c0c2971a9b55b3af7c8ad935edc639577048b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2472823
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>