Reland "[Autotest] Migrating client/bin/job.py to python3"
This is a reland of 96cdb58d2e41d7aae5bd276ba2aa48eaba12e791
Fixed the issue which resulted in a revert.

TEST=local testing with custom .py's. Unfortunantly, with this fix, and
with the old ToT (ie, prior to the bad CL landing), it fails when I
forced it to rebuild the deps on "Unhandled KeyError: 'SYSROOT'".
I am guessing this is a var set within the lxc/drone.
However, the custom testing I did showed the issue and that this will
fix it, and be py3 compatible.

Original change's description:
> [Autotest] Migrating client/bin/job.py to python3
>
> This is causing issues. Making the file solo to help debug CQ problems
>
> BUG=chromium:990593
> TEST= py_compile in py2 and py3. Applicable unittests. CQ. dummy_Pass
>
> Change-Id: I90856bc5652965770fc8300423bba6c61fc21548
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2424744
> Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
> Commit-Queue: Derek Beckett <dbeckett@chromium.org>
> Tested-by: Derek Beckett <dbeckett@chromium.org>
> Auto-Submit: Derek Beckett <dbeckett@chromium.org>

Bug: chromium:990593
Change-Id: I22c0703b55341e08236413d56a704149242cd075
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2450830
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>