Put name of in-progress test in crash metadata.
When we start uploading crashes from test results, it will be useful to
know during which test a crash occurred (if any).

Also, run git cl format.

This is a reland of https://crrev.com/c/2429928, fixed to work for
server tests and cheets tests, and to be more robust.

BUG=chromium:1119443
TEST=test_that $DUT logging_UserCrash logging_GenerateCrashFiles \
     cheets_CTS_P.9.0_r13.x86.wm-presubmit

Change-Id: I7fab8f4567a7318e9a9aea64b34df5d669d49ea3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2481972
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>