cheets_CTS/GTS: Force laptop mode on Hatch.
Hatch.Kohaku in the lab seems to be physically set up in
the tablet mode. This is not the default mode we want to
test because the device is primarily a PC laptop.

This CL enforces the mode before starting the test.

BUG=b:169373919
TEST=tail -f /var/log/power_manager/powerd.LATEST | grep laptop

Change-Id: I8f7bc51e3e931cc8dbb669982484f087f6220ec4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2430532
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Jiyoun Ha <jiyounha@chromium.org>
Reviewed-by: Jiyoun Ha <jiyounha@chromium.org>