bluetooth: Add check to test_register_advertisement for tx_power
Now that our clients can specify Tx Power in advertisements, we want to
test this feature. If the advertisement used in the test has a specified
tx power, this new check verifies that it is properly sent over both
MGMT and HCI interfaces to controller, and that the selected TX power by
the controller is propagated back up to the client.

BUG=b:152433556
TEST=Ran AdvSanity on hatch and kukui

Change-Id: I4b0ce7cb14b3b4ab570569362ec5774d65cce58b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2702454
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Manish Mandlik <mmandlik@chromium.org>