network_WiFi_RoamFT: add ping test
Add a ping test from the router to check layer 3 connectivity after a
roam. We don't do the standard ping check, which pings the router from
the client, because the bridge setup used in this test prevents the
router from responding to pings.

BUG=chromium:1070321
TEST=network_WiFi_RoamFT.*, tried pinging from other bridge interface
and it didn't work.

Change-Id: Ic107d0b7e38823a61f6857ec86b51413ce519a4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2433165
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>