firmware_test: Set needs_servo_usb False by default.
Since FirmwareTest is used by more than just firmware_* tests, this
commit changes the default value to be False instead of True, and
specifically marks True on the tests that are known to need it.

To make the variable easier to find, the name is now capitalized.

BUG=b:169570370
TEST=Run test with a device that has bad servo_usb_state.

Change-Id: I55a759d6e10b8f30021499475e22917543b6b570
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2434634
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>