platform_ServoPowerStateController: move `snk` to after `:wqpoweroff`
This change moves switching the role on the servo v4 from `src` to `snk`
after the DUT has been turned off. This is to address an issue on grunt
(so far) where if the DUT is being asked to go into S5 and subsequently
the v4 to change from `src` to `snk`, the device can have issue exposing
the SBU lines from the cr50 to the v4/labstation, and ccd_cr50
communication fails. This state then cannot be recovered through the
servo (as the communication fails) and the DUT becomes unrecoverable.

We defensively insert ten second sleeps and a ->src setting before the
->snk setting to make this as robust as possible in the lab. Empirically,
it prevents grunts from entering a bad state.

BUG=b:169797292
TEST=Test on zorks and grunts in the lab

Change-Id: Idaa53659c997ab7022dfb062784d265d7f134195
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2442036
Tested-by: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>