firmware_ECCharging: Don't test when DUT performs discharge on AC
Some boards (eg. eve) performs discharge on AC when battery reports
that it is fully charged and doesn't want more charge. This can lead to
situation in which charge is below 100% and battery is still discharging,
for example:

> battery
  Status:    0x00e0 FULL DCHG INIT
  Param flags:00000002
  Temp:      0x0baa = 298.6 K (25.5 C)
  V:         0x218d = 8589 mV
  V-desired: 0x0000 = 0 mV
  I:         0xff22 = -222 mA(DISCHG)
  I-desired: 0x0000 = 0 mA
  Charging:  Not Allowed
  Charge:    98 %

> chgstate
  state = discharge
  ac = 1
  batt_is_charging = 0

Test expects that battery is being charged when it is below 100%,
which is not true during discharge on AC.

This patch adds checking if discharge on AC takes place, if so just
exit from test with appropriate message.

BUG=b:162254118
TEST=Charge battery to 100%, this should trigger discharge on AC (check
     it using 'battery' and 'chgstate' commands).
     Run firmware_ECCharging test. It should report that DUT performs
     discharge on AC and finish successfully.

Change-Id: Ic42a55cf98082984b95ea0d7d5daba83755d34bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2440896
Tested-by: Patryk Duda <pdk@semihalf.com>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>