bluetooth: Add functionality to flag system failures in BT tests
Some failures, such as USB disconnects or bluetoothd crashes will cause
tests to fail, and require us to manually investigate the logs to
identify these root causes. This change wraps all QuickSanity tests in a
log checker, such that if a common failure is detected in system logs,
it is appended to the failure reason reported to stainless. This should
imrove our ability to correlate these root causes to the reported error
in an automated way.

BUG=b:170221594
TEST=QuickSanity and AdvSanity on Hatch

Change-Id: If0342598129752bf08de47dc6974092dd8afccbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2454191
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>