bluetooth: a2dp write to raw file
Let the btpeer record the A2DP audio streams to a raw file
directly instead of piping to another process to avoid the
little/big endian issue ocurring during the piping process.

The raw file, compared to a wav file, is easier to divide
into chunks for verification.

In addition, remote copying the chunk file to the DUT is
performed internally inside the btpeer to be more efficient.

BUG=b:169039281
TEST=Conduct both the A2DP test and A2DP long test on DUTs
in the lab.

Change-Id: I1a59ba5181e98a3da6b259f1d29f9d211083650e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2461187
Reviewed-by: Michael Sun <michaelfsun@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>