hardware_StorageFioOther: pass through the blkdiscard argument
fio_test has an argument to specify whether to do a blkdiscard
before running the fio test. We need to be able to specify and pass
this argument through hardware_StorageFioOther.

Not being able to pass through blkdiscard=False is what is causing
the verify_only test to fail in hardware_StorageSuspendStress since
we are doing a blkdiscard by default before we try to verify the data
written.

BUG=chromium:1131281
TEST=Ran hardware_StorageSuspendStress.test_other with an sd card

Change-Id: I347e4ec0de5bd9e2168a123fb189ec3660c327ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2459130
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>