fingerprint: Disable biod's upstart job
On some platforms (e.g. Zork), an AP reboot is executed after running
flash_fp_mcu. A side effect of this AP reboot is biod will initialize
the entropy. This can mess up with test's treatment or expectation of
the entropy. Disable biod's upstart job for the duration of the test
so that we are not affected by biod after reboot.

Also move the treatment for fp_updater together for readability.

BUG=b:170352568
TEST=firmware_Fingerprint.AddEntropy,
     firmware_Fingerprint.ROCanUpdateRW
     => Pass on 13533.0.0

Change-Id: I1c20d955f5f3f03e46661d895d4aeac36c35189a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2462707
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>