FWUpdateThenSleep: Remove dependency on setup_pdtester
setup_pdtester is primarily used to setup for pd_faft. It is used in
this test to ensure the battery is sufficiently charged.

One of the steps in setup_pdtester is causing servo_v4 to fail when
a servo_micro is not connected to the servo_v4 like it is during pd_faft.

Replace setup_pdtester with a call to put_host_battery_in_range to avoid
the unnecessary dependency.

BUG=b:170343710
TEST=Pass test on Zork

Change-Id: I139d77317e4c9aa8999dc23f05e88386b53af6c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2462871
Commit-Queue: Rob Barnes <robbarnes@google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Tested-by: Rob Barnes <robbarnes@google.com>
Auto-Submit: Rob Barnes <robbarnes@google.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Philip Chen <philipchen@chromium.org>