bluetooth: Catch TestError and mark as error
Catch error.TestError in the quicktest testcase wrapper and emit it as
a failure in the suite. This makes sure that even if an individual test
case emits a TestError, the suite continues running and eventually
generates a summary.

BUG=b:169135684
TEST=Modifed AUSanity test to always emit TestError and run suite. Observe that all tests run and report is generated.

Change-Id: I6ea8ad233aa7c99fdb07cb261063715f8596cec2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2464142
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Auto-Submit: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>