FAFT: Add check to ECWakeSource test for devices in tablet mode
Modified firmware_ECWakeSource test to check for devices in tablet mode
using the EC ksstate command. When a device is in tablet mode the test
will check the keyboard does not wake the device, if the device is in
normal mode the existing check for wake on keypress is done.

BUG=b:158130073
TEST=Verified on local kohaku not in tablet mode and verified on lab DUT
     in tablet mode.

Change-Id: I951d4ad18655d8e7d8feafa4e016b48895ecb26a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2466153
Tested-by: Brent Peterson <brentpeterson@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Brent Peterson <brentpeterson@chromium.org>