mode_switcher: workaround the bug in RO with USB
There is a well known bug that DUT does not see USB stick in the
recovery mode during recovery tests. The root cause of that is an
issue in EC RO firmware - it does not supply Vbus in SRC_ACCESSORY
(DUT boots to recovery as SRC due to the assumption of no PD in RO).
It has been fixed in 251212fb(removes ACCESSORY modes), but since we
don't update RO it has to be workarounded in tests.

Check if Servo is snk and if Vbus isn't supplied, then the issue
occurs. In that case, change Servo PR to supply Vbus correctly and
perform PD Data Swap(PD should be available because most boards used
for testing purposes have WP disabled).

If Servo is already SRC and EC if UFP(case after EC reboot), perform
PD Data Swap as well. This can be done in the workaround placed in
servod, but it may be removed in some time so make sure everything
is correct.

The workaround is used only for boards which have the faft_config
ec_ro_vbus_bug to avoid hiding potentially new bugs in future devices.

BUG=b:159938441
TEST=Run recovery tests

Change-Id: Ia0d115bb46f21b3c38b3d2d0b6a22d6b61a9ffa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2475053
Tested-by: Dawid Niedźwiecki <dn@semihalf.com>
Commit-Queue: Dawid Niedźwiecki <dn@semihalf.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Patryk Duda <pdk@semihalf.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Greg Edelston <gredelston@google.com>