autotest: Remove network_WiFi_ScanPerformance.
As its Tast counterpart, tast.wifi.ScanPerf, was stable and promoted to
suite:wifi_perf for months. It is time to retire the Autotest
counterpart.

Also, since it is a performance test, the Tast version does not run in
suite:wifi_matfunc and suite:wificell-cq.

BUG=b:159092883, b:156249427
TEST=none

Cq-Depend: chromium:2477133
Change-Id: I4989de29e3e04b5df5d8a8624e96b537c774a439
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2476355
Tested-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Anthony Hinton <hinton@google.com>
Commit-Queue: Shuo-Peng Liao <deanliao@chromium.org>