fingerprint: Move RDP0 test from bash to autotest (python)
On Zork the RDP0 test requires an AP reboot. Move the test logic
from bash script to the autotest class itself, so that we can
reboot AP in the middle of the test.

The same thing has been done for RDP1 in http://crrev/c/2464244

BUG=b:170739632
TEST=suite:fingerprint passes

Change-Id: Iac34873772eac41ef4d00907129a32b1f5ea7038
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2477884
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>