servo: _WrapServoErrors: Move error message to debug
No control named x error is raised as an exception.
If caller catches the exception it was not an error, this is causing
error log spam. If caller does not catch the exception it will be logged
as an error in the normal way.

BUG=None
TEST=test_that firmware_FAFTSetup

Change-Id: I54307b2134cd40cfa426483be87c2cae2b9ac255
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2476356
Reviewed-by: Brent Peterson <brentpeterson@chromium.org>
Commit-Queue: Evan Benn <evanbenn@chromium.org>
Tested-by: Evan Benn <evanbenn@chromium.org>