platform_ServoPowerStateController: Detect EC RO bug
There is a bug in EC/RO(b/159938441), which causes not suppling
Vbus(in recovery), thus USB Stick is not detected.
With this bug present, DUT won't boot from USB(without additional
PD operations, based on FAFT config which is not allowed in this test),
so omit checking "power off" in recovery mode.

BUG=b:170894065
TEST=Run platform_ServoPowerStateController.USBPluggedin

Change-Id: Ied2117ed6253ade3c5e93ea8c3cfac8029f7389d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2510149
Tested-by: Dawid Niedźwiecki <dn@semihalf.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>